chore: move ad card to it's own structure
