+++ packages/shared/src/components/cards/ad/common/AdPixel.tsx
@@ -1,0 +1,18 @@
+import React, { ReactElement } from 'react';
+import type { Ad } from '../../../../graphql/posts';
+
+export const AdPixel = ({ pixel }: Pick<Ad, 'pixel'>): ReactElement => {
+  return (
+    <>
+      {pixel?.map((p) => (
+        <img
+          src={p}
+          key={p}
+          data-testid="pixel"
+          className="hidden size-0"
+          alt="Pixel"
+        />
+      ))}
+    </>
+  );
+};
